centralize the etl functions
